update
- fix calculator

+++ app/src/main/java/com/training/app/contract/CalculatorContract.java
@@ -2,6 +2,8 @@ package com.training.app.contract;
 
 import com.training.app.object.Calculator;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Dell on 7/4/2017.
  */
@@ -13,7 +15,7 @@ public interface CalculatorContract {
     }
 
     interface View {
-        void updateDisplay(double value);
+        void updateDisplay(BigDecimal value);
     }
 
 }
